minor refactor to maybe_get_vit_flash_attn_backend()
